+++ src/inventorum.ebay/inventorum/ebay/apps/products/tests/test_serializers.py
@@ -64,7 +64,7 @@ class TestEbayProductSerializer(UnitTestCase, ShippingServiceConfigurableSeriali
         return CategoryFactory.create(name="Some category", parent=some_parent)
 
     def get_default_product(self):
-        product = EbayProductFactory.create(inv_id=1337,
+        product = EbayProductFactory.create(inv_id=1225146588560351744L,
                                             category=self.get_default_category())
         product.shipping_services.create(service=self.get_shipping_service_dhl(),
                                          cost=D("5.00"),
@@ -79,7 +79,7 @@ class TestEbayProductSerializer(UnitTestCase, ShippingServiceConfigurableSeriali
         subject = EbayProductSerializer(product)
         self.assertEqual(subject.data, {
             "id": product.id,
-            "inv_id": 1337,
+            "inv_id": "1225146588560351744",
             "listing_url": None,
             "is_published": False,
             "is_click_and_collect": False,