+++ CHANGES.rst
@@ -2,9 +2,18 @@
 CHANGES
 =======
 
-develop
-=======
-2015-06-16 0.1.6 
+2015-06-25 0.1.8
+================
+- Use `product_inv_id` for published product lookup in core products sync
+
+2015-06-19 0.1.7
+================
+- Change inv_id in serializers to string format
+- Adapt core image serializer to the new image format
+- INV-4695 Use core inv_id in ebay service
+- INV-4631/INV-4457 Avoid double publishing by using database locks in the resource
+
+2015-06-16 0.1.6
 ================
 - add ``rc`` environment for release candidate environment
 - add db configs for ``aero.inventorum.net``